:zap: 优化代码

+++ src/main/java/cn/taketoday/context/cglib/proxy/CallbackHelper.java
@@ -92,7 +92,7 @@ public abstract class CallbackHelper implements CallbackFilter {
   }
 
   public int accept(Method method) {
-    return methodMap.get(method).intValue();
+    return methodMap.get(method);
   }
 
   public int hashCode() {
